prompt_tuning_attack/requirements.txt
torch
transformers
datasets
Pillow
numpy
matplotlib
opencv-python
scikit-image
tqdm